requirements.txt
langchain
langchain-community
openai
chromadb
Langchain_openai
langchain_chroma
python-dotenv
spacy
